requirements.txt
fastapi
uvicorn
pydantic
python-multipart
stripe
mysql-connector-python
pytz
shutil
typing
python-dotenv
